Tighten HomeKit readability refactor
