--- src/homekit/blinds.rs
@@ -1,42 +1,0 @@
-//! Somfy blind accessory metadata shared by the HomeKit adapter.
-
-use crate::gpio::Channel;
-
-#[derive(Copy, Clone, Debug)]
-pub struct Blind {
-    pub aid: u64,
-    pub name: &'static str,
-    pub channel: Channel,
-    pub serial: &'static str,
-}
-
-pub const BLINDS: &[Blind] = &[
-    Blind {
-        aid: 2,
-        name: "Blind 1",
-        channel: Channel::L1,
-        serial: "somfy-L1",
-    },
-    Blind {
-        aid: 3,
-        name: "Blind 2",
-        channel: Channel::L2,
-        serial: "somfy-L2",
-    },
-    Blind {
-        aid: 4,
-        name: "Blind 3",
-        channel: Channel::L3,
-        serial: "somfy-L3",
-    },
-    Blind {
-        aid: 5,
-        name: "Blind 4",
-        channel: Channel::L4,
-        serial: "somfy-L4",
-    },
-];
-
-pub fn find_blind(aid: u64) -> Option<&'static Blind> {
-    BLINDS.iter().find(|b| b.aid == aid)
-}
