+++ src/homekit/mod.rs
@@ -10,11 +10,9 @@ use crate::hap::state::{FileHapStore, HapState};
 use crate::remote::RemoteControl;
 
 mod accessory_db;
-mod blinds;
 pub mod config;
 mod position_cache;
 pub mod positions;
-mod reads;
 pub mod somfy;
 mod target_writes;
