+++ src/homekit/position_cache.rs
@@ -8,9 +8,47 @@ use crate::hap::runtime::{CharacteristicEvent, CharacteristicId};
 use crate::homekit::accessory_db::{
     IID_CURRENT_POSITION, IID_POSITION_STATE, IID_TARGET_POSITION, POSITION_STATE_STOPPED,
 };
-use crate::homekit::blinds::{find_blind, Blind, BLINDS};
 use crate::homekit::positions;
 use serde_json::json;
+
+#[derive(Copy, Clone, Debug)]
+pub struct Blind {
+    pub aid: u64,
+    pub name: &'static str,
+    pub channel: Channel,
+    pub serial: &'static str,
+}
+
+pub const BLINDS: &[Blind] = &[
+    Blind {
+        aid: 2,
+        name: "Blind 1",
+        channel: Channel::L1,
+        serial: "somfy-L1",
+    },
+    Blind {
+        aid: 3,
+        name: "Blind 2",
+        channel: Channel::L2,
+        serial: "somfy-L2",
+    },
+    Blind {
+        aid: 4,
+        name: "Blind 3",
+        channel: Channel::L3,
+        serial: "somfy-L3",
+    },
+    Blind {
+        aid: 5,
+        name: "Blind 4",
+        channel: Channel::L4,
+        serial: "somfy-L4",
+    },
+];
+
+pub fn find_blind(aid: u64) -> Option<&'static Blind> {
+    BLINDS.iter().find(|b| b.aid == aid)
+}
 
 /// HomeKit blind position after snapping a 0–100 request to an endpoint.
 #[derive(Copy, Clone, Debug, PartialEq, Eq)]
