+++ src/server.rs
@@ -282,92 +282,69 @@ async fn websocket(stream: WebSocket, state: Arc<AppState>, client_name: String,
 mod tests {
     use super::*;
 
-    fn validated(command: &str, channel: Option<Channel>, long: bool) -> ValidatedCommandRequest {
-        CommandRequest {
+    fn validate(
+        command: &str,
+        channel: Option<Channel>,
+        long: bool,
+    ) -> Result<(Command, Option<Channel>), String> {
+        let request = CommandRequest {
             command: command.to_string(),
             channel,
             long,
         }
-        .validate()
-        .unwrap()
+        .validate()?;
+        Ok((request.command, request.channel))
     }
 
     #[test]
     fn validate_accepts_select_with_channel() {
         assert_eq!(
-            validated("select", Some(Channel::L2), false),
-            ValidatedCommandRequest {
-                command: Command::Select,
-                channel: Some(Channel::L2),
-            }
+            validate("select", Some(Channel::L2), false).unwrap(),
+            (Command::Select, Some(Channel::L2))
         );
     }
 
     #[test]
     fn validate_accepts_select_without_channel() {
         assert_eq!(
-            validated("select", None, false),
-            ValidatedCommandRequest {
-                command: Command::Select,
-                channel: None,
-            }
+            validate("select", None, false).unwrap(),
+            (Command::Select, None)
         );
     }
 
     #[test]
     fn validate_accepts_directional_channel() {
         assert_eq!(
-            validated("up", Some(Channel::L1), false),
-            ValidatedCommandRequest {
-                command: Command::Up,
-                channel: Some(Channel::L1),
-            }
+            validate("up", Some(Channel::L1), false).unwrap(),
+            (Command::Up, Some(Channel::L1))
         );
     }
 
     #[test]
     fn validate_rejects_prog_without_channel() {
-        let err = CommandRequest {
-            command: "prog".into(),
-            channel: None,
-            long: false,
-        }
-        .validate()
-        .unwrap_err();
+        let err = validate("prog", None, false).unwrap_err();
         assert!(err.contains("requires a channel"));
     }
 
     #[test]
     fn validate_accepts_prog_with_channel() {
         assert_eq!(
-            validated("prog", Some(Channel::L1), false),
-            ValidatedCommandRequest {
-                command: Command::Prog,
-                channel: Some(Channel::L1),
-            }
+            validate("prog", Some(Channel::L1), false).unwrap(),
+            (Command::Prog, Some(Channel::L1))
         );
     }
 
     #[test]
     fn validate_promotes_prog_with_long_to_prog_long() {
         assert_eq!(
-            validated("prog", Some(Channel::L1), true),
-            ValidatedCommandRequest {
-                command: Command::ProgLong,
-                channel: Some(Channel::L1),
-            }
+            validate("prog", Some(Channel::L1), true).unwrap(),
+            (Command::ProgLong, Some(Channel::L1))
         );
     }
 
     #[test]
     fn validate_rejects_long_on_non_prog_commands() {
-        let err = CommandRequest {
-            command: "up".into(),
-            channel: Some(Channel::L1),
-            long: true,
-        }
-        .validate()
-        .unwrap_err();
+        let err = validate("up", Some(Channel::L1), true).unwrap_err();
         assert!(err.contains("only valid with prog"));
     }
